+++ libnail/CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - now stores relevant timings
     - refactored AlignmentBuilder for changes
 - DpMatrixSparse should now resize properly
+- CloudSearchScores renamed to CloudSearchResults, added num_cells_computed field
 
 ### Deprecated